moved time controls into their own directory
